requirements.txt
gradio
sentence-transformers
faiss-cpu
numpy
transformers
torch
sentencepiece
einops
accelerate